refactor(channel): extract thread keymap + process lifecycle
Phase 2 continuation. Two extractions:

(1) channel_thread_input.go (124 LOC, new): updateThread keymap.
    Mirrors the main composer keymap (motion, history, line insert,
    scroll) but writes through to threadInput / threadInputPos /
    threadScroll. Named *_thread_input.go because channel_thread.go is
    already taken by the thread-panel renderer (renderThreadPanel /
    renderThreadInput).

(2) channel_lifecycle.go (102 LOC, new): tickChannel,
    killTeamSession, runChannelView, reportChannelCrash. Process /
    lifecycle helpers — best-effort cleanup at exit, recover-wrapped
    Bubbletea program entrypoint, last-words-on-stderr crash handler.

  - cmd/wuphf/channel.go: -185 LOC (2300 -> 2115)
  - cmd/wuphf/channel_thread_input.go: +124 LOC (new)
  - cmd/wuphf/channel_lifecycle.go: +102 LOC (new)

Cumulative channel.go reduction across Phase 2:
  4807 -> 2115
  Net: -2692 LOC; channel.go shrunk by 56%.

The remaining bulk is the Update method (~1500 LOC). Splitting Update
cleanly requires either grouping cases by message-type family or
introducing a dispatch table — both larger architectural moves than
this session is doing. Channel.go below 1500 stays a follow-up target.

No behavior change. cmd/wuphf tests pass.
