+++ cmd/wuphf/channel.go
@@ -9,13 +9,12 @@ import (
 	"io"
 	"net/http"
 	"os"
-	"os/exec"
-	"runtime/debug"
 	"strconv"
 	"strings"
 	"time"
 
 	tea "github.com/charmbracelet/bubbletea"
+	"github.com/charmbracelet/lipgloss"
 
 	"github.com/nex-crm/wuphf/cmd/wuphf/channelui"
 	"github.com/nex-crm/wuphf/internal/brokeraddr"
@@ -2305,115 +2304,6 @@ func (m channelModel) mainPanelGeometry(mainW, contentH int) (headerH, msgH int,
 	return headerH, msgH, popupRows
 }
 
-// updateThread handles key events when the thread panel is focused.
-func (m channelModel) updateThread(msg tea.KeyMsg) (tea.Model, tea.Cmd) {
-	if motionKey, ok := composerMotionKey(msg); ok {
-		if nextPos, handled := channelui.MoveComposerCursor(m.threadInput, m.threadInputPos, motionKey); handled {
-			m.threadInputPos = nextPos
-			m.updateThreadOverlays()
-		}
-		return m, nil
-	}
-	key := msg.String()
-	if msg.Type == tea.KeyCtrlJ {
-		key = "ctrl+j"
-	}
-	switch key {
-	case "enter":
-		if len(m.threadInput) > 0 {
-			text := string(m.threadInput)
-			trimmed := strings.TrimSpace(text)
-			if strings.HasPrefix(trimmed, "/") {
-				m.threadInputHistory.Record(m.threadInput, m.threadInputPos)
-				return m.runCommand(trimmed, m.threadPanelID)
-			}
-			m.threadInputHistory.Record(m.threadInput, m.threadInputPos)
-			m.threadInput = nil
-			m.threadInputPos = 0
-			m.posting = true
-			return m, postToChannel(text, m.threadPanelID, m.activeChannel)
-		}
-	case "backspace":
-		if m.threadInputPos > 0 {
-			m.threadInputHistory.ResetRecall()
-			m.threadInput = append(m.threadInput[:m.threadInputPos-1], m.threadInput[m.threadInputPos:]...)
-			m.threadInputPos--
-			m.updateThreadOverlays()
-		}
-	case "ctrl+u":
-		m.threadInputHistory.ResetRecall()
-		m.threadInput = nil
-		m.threadInputPos = 0
-		m.updateThreadOverlays()
-	case "ctrl+p":
-		if snapshot, ok := m.threadInputHistory.Previous(m.threadInput, m.threadInputPos); ok {
-			m.restoreThreadSnapshot(snapshot)
-		}
-	case "ctrl+n":
-		if snapshot, ok := m.threadInputHistory.Next(); ok {
-			m.restoreThreadSnapshot(snapshot)
-		}
-	case "ctrl+a":
-		m.threadInputPos = 0
-		m.updateThreadOverlays()
-	case "ctrl+e":
-		m.threadInputPos = len(m.threadInput)
-		m.updateThreadOverlays()
-	case "ctrl+j":
-		m.threadInputHistory.ResetRecall()
-		ch := []rune{'\n'}
-		tail := make([]rune, len(m.threadInput[m.threadInputPos:]))
-		copy(tail, m.threadInput[m.threadInputPos:])
-		m.threadInput = append(m.threadInput[:m.threadInputPos], append(ch, tail...)...)
-		m.threadInputPos++
-		m.updateThreadOverlays()
-	case "left":
-		if m.threadInputPos > 0 {
-			m.threadInputPos--
-			m.updateThreadOverlays()
-		}
-	case "right":
-		if m.threadInputPos < len(m.threadInput) {
-			m.threadInputPos++
-			m.updateThreadOverlays()
-		}
-	case "up":
-		m.threadScroll++
-	case "down":
-		m.threadScroll--
-		if m.threadScroll < 0 {
-			m.threadScroll = 0
-		}
-	case "pgup":
-		m.threadScroll += 5
-	case "pgdown":
-		m.threadScroll -= 5
-		if m.threadScroll < 0 {
-			m.threadScroll = 0
-		}
-	default:
-		if ch := composerInsertRunes(msg); len(ch) > 0 {
-			m.threadInputHistory.ResetRecall()
-			m.threadInput, m.threadInputPos = channelui.InsertComposerRunes(m.threadInput, m.threadInputPos, ch)
-			m.updateThreadOverlays()
-		} else if len(msg.String()) == 1 || msg.Type == tea.KeyRunes {
-			ch := msg.Runes
-			if len(ch) == 0 {
-				ch = []rune(msg.String())
-			}
-			if len(ch) > 0 {
-				m.threadInputHistory.ResetRecall()
-				tail := make([]rune, len(m.threadInput[m.threadInputPos:]))
-				copy(tail, m.threadInput[m.threadInputPos:])
-				m.threadInput = append(m.threadInput[:m.threadInputPos], append(ch, tail...)...)
-				m.threadInputPos += len(ch)
-				m.updateThreadOverlays()
-			}
-		}
-	}
-	return m, nil
-}
-
 // updateSidebar handles key events when the sidebar is focused.
 func (m channelModel) updateSidebar(msg tea.KeyMsg) (tea.Model, tea.Cmd) {
 	roster := channelui.MergeOfficeMembers(m.officeMembers, m.members, m.currentChannelInfo())
@@ -3985,76 +3875,4 @@ func applyProviderSelection(providerName string) tea.Cmd {
 	}
 }
 
-func tickChannel() tea.Cmd {
-	return tea.Tick(2*time.Second, func(t time.Time) tea.Msg {
-		return channelTickMsg(t)
-	})
-}
-
-// killTeamSession kills the entire wuphf-team tmux session and all agent processes.
-func killTeamSession() {
-	// Best-effort cleanup at process exit; cap each step so a hung tmux or
-	// broker doesn't keep us alive forever.
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-	// Kill tmux session (kills all agent processes in all panes/windows)
-	_ = exec.CommandContext(ctx, "tmux", "-L", "wuphf", "kill-session", "-t", "wuphf-team").Run()
-	// Stop the broker
-	req, err := http.NewRequestWithContext(ctx, http.MethodGet, brokerURL("/health"), nil)
-	if err != nil {
-		return
-	}
-	if resp, err := http.DefaultClient.Do(req); err == nil {
-		_ = resp.Body.Close()
-	}
-}
-
-func runChannelView(threadsCollapsed bool, initialApp channelui.OfficeApp, skipSplash bool) {
-	defer func() {
-		if r := recover(); r != nil {
-			reportChannelCrash(fmt.Sprintf("panic: %v\n\n%s", r, debug.Stack()))
-		}
-	}()
-
-	// Check if onboarding is needed before launching the channel view.
-	if os.Getenv("WUPHF_SKIP_ONBOARDING") == "" {
-		state, err := fetchOnboardingState(brokerBaseURL())
-		if err == nil && !state.Onboarded {
-			om := newOnboardingModel(brokerBaseURL(), 0, 0)
-			op := tea.NewProgram(om, tea.WithAltScreen())
-			if _, err := op.Run(); err != nil {
-				reportChannelCrash(fmt.Sprintf("onboarding error: %v\n", err))
-				return
-			}
-			// Fall through to channel view after onboarding completes.
-		}
-	}
-
-	if !skipSplash && os.Getenv("WUPHF_NO_SPLASH") == "" {
-		splash := tea.NewProgram(newSplashModel(), tea.WithAltScreen())
-		if _, err := splash.Run(); err != nil {
-			reportChannelCrash(fmt.Sprintf("splash error: %v\n", err))
-			return
-		}
-	}
-
-	p := tea.NewProgram(newChannelModelWithApp(threadsCollapsed, initialApp), tea.WithAltScreen())
-	if _, err := p.Run(); err != nil {
-		reportChannelCrash(fmt.Sprintf("channel view error: %v\n", err))
-	}
-}
-
-func reportChannelCrash(details string) {
-	_ = channelui.AppendChannelCrashLog(details)
-	fmt.Fprintln(os.Stderr, "WUPHF channel crashed.")
-	fmt.Fprintln(os.Stderr, "Log:", channelui.ChannelCrashLogPath())
-	fmt.Fprintln(os.Stderr, "")
-	fmt.Fprintln(os.Stderr, "The rest of the team is still running.")
-	if strings.TrimSpace(os.Getenv("WUPHF_HEADLESS_PROVIDER")) != "" {
-		fmt.Fprintln(os.Stderr, "Restart WUPHF when ready to reconnect to the headless office runtime.")
-	} else {
-		fmt.Fprintln(os.Stderr, "Use `tmux -L wuphf attach -t wuphf-team` to inspect panes,")
-		fmt.Fprintln(os.Stderr, "then restart WUPHF when ready.")
-	}
-	select {}
-}
+// Process lifecycle helpers moved to channel_lifecycle.go.
